scripts/requirements.txt
fastapi==0.115.4
numpy==2.1.3
openai==1.54.2
psycopg2-binary==2.9.10
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.36
uvicorn==0.32.0